chore: Move setup script to a repo package
