+++ internals/setup-packages/src/hoist-peer-dependencies.ts
