+++ internals/setup-packages/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@repo/tsconfig",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["src"],
+  "exclude": ["src/**/*.test.ts"]
+}
