+++ package.json
@@ -58,7 +58,7 @@
     "prepare": "husky && turbo tsc attw",
     "prepack": "find packages/* -maxdepth 1 -type f -name 'README*' -exec sed -i '' -e 's/utm_source=github/utm_source=npmjs/g' {} + && find packages/* -maxdepth 1  -type f -name package.json -exec sh -c 'jq \".devDependencies |= with_entries(select(.value != \\\"workspace:*\\\"))\" \"$1\" > \"$1.tmp\" && mv \"$1.tmp\" \"$1\"' _ {} \\;",
     "postpack": "git restore packages/*/package.json packages/*/README.md",
-    "setup": "node ./tools/setup-packages.ts && prettier --write . --log-level=silent",
+    "setup": "setup-packages && prettier --write . --log-level=silent",
     "pretest": "yarn tsc && oxlint && prettier --check . && eslint .",
     "test": "vitest --run packages && turbo test",
     "tsc": "turbo tsc && tsc -p tsconfig.json"
@@ -72,10 +72,8 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.38.0",
-    "@repo/hoist-peer-dependencies": "workspace:*",
     "@repo/tsconfig": "workspace:*",
     "@types/node": "^24.9.1",
-    "@types/semver": "^7",
     "@vitest/coverage-v8": "^3.2.4",
     "@vitest/ui": "^3.2.4",
     "eslint": "^9.38.0",
@@ -83,14 +81,13 @@
     "eslint-plugin-n": "^17.23.1",
     "eslint-plugin-oxlint": "^1.7.0",
     "globals": "^16.4.0",
-    "globby": "^15.0.0",
     "husky": "^9.1.7",
     "isolate-monorepo-package": "workspace:*",
     "lerna": "^9.0.0",
     "nano-staged": "^0.8.0",
     "oxlint": "^1.7.0",
     "prettier": "^3.6.2",
-    "semver": "^7.7.3",
+    "setup-packages": "workspace:*",
     "turbo": "^2.5.8",
     "type-fest": "^5.1.0",
     "typescript": "^5.9.3",
